% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IMData}
\alias{IMData}
\title{IMaGES Test Data}
\format{Objects of class \code{"list"}.}
\usage{
data(IMData)
}
\description{
Test data to be used with IMaGES. Included is
a list called \code{IMData}, a list containing three
data files that can be passed into IMaGES.
}
\examples{
data(IMData)

result <- IMaGES(matrices=IMData)

plotAll(result)



}
\keyword{data}
