% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IMTrue}
\alias{IMTrue}
\title{IMaGES Test Data}
\format{Objects of class \code{"list"}.}
\usage{
data(IMData)
}
\description{
Test data to be used with IMaGES. Included is
a list called \code{IMTrue}, a list containing the original
graph structures for each dataset in \code{IMTrue}.
}
\examples{
data(IMData)
data(IMTrue)

result <- IMaGES(matrices=IMData)

plotAll(result)

par(mfrow=c(1,length(IMTrue)))

for (i in 1:length(IMTrue)) {
  graph::plot(IMTrue[[i]])
}


}
\keyword{data}
